# app/Http/Requests/Post/PostChangeStatusRequest.php
<?php

namespace App\Http\Requests\Post;

use Illuminate\Foundation\Http\FormRequest;
use OpenApi\Annotations as OA;

/**
 * @OA\Schema(
 *      required={"status"},
 *      schema="PostChangeStatusRequest",
 *      @OA\Property(property="status", type="int", example="0", description="Post status"),
 * )
 */
class PostChangeStatusRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
     */
    public function rules(): array
    {
        return [
            'status' => 'required|integer|between:0,1',
        ];
    }
}