# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Service 'Short blog API'",
        "version": "0.1"
    },
    "paths": {
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "List of posts",
                "description": "List of posts",
                "operationId": "/api/posts(GET)",
                "parameters": [
                    {
                        "name": "sort_column",
                        "in": "query",
                        "description": "Sort column",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Sort direction",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Posts per page",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "keys",
                        "in": "query",
                        "description": "Post keys",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "",
                            "example": ""
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name and content",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    },
                    {
                        "name": "published_gte",
                        "in": "query",
                        "description": "Publish date is greater than",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    },
                    {
                        "name": "published_lte",
                        "in": "query",
                        "description": "Publish date less than",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status",
                        "required": false,
                        "schema": {
                            "type": "int",
                            "example": ""
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Includes",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "author"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return list of posts",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create post",
                "operationId": "/api/posts(POST)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return model of new post",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get post by id",
                "description": "Show the specified resource from storage.",
                "operationId": "/api/posts/{id}(GET)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": ""
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Includes",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "author"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get exists post"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update post",
                "operationId": "/api/posts/{id}(PUT)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return post model",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete post",
                "description": "Remove the specified resource from storage.",
                "operationId": "/api/posts/{id}(DELETE)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete exists post",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                }
            }
        },
        "/api/post-change-status/{id}": {
            "patch": {
                "tags": [
                    "Posts"
                ],
                "summary": "Change post status",
                "description": "Change status",
                "operationId": "/api/post-change-status/{id}(PATCH)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": ""
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostChangeStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Change post status",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PostChangeStatusRequest": {
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "description": "Post status",
                        "type": "integer",
                        "example": "0"
                    }
                },
                "type": "object"
            },
            "PostCreateRequest": {
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "description": "Post title",
                        "type": "string",
                        "example": ""
                    },
                    "content": {
                        "description": "Post content",
                        "type": "string",
                        "example": ""
                    },
                    "status": {
                        "description": "Post status",
                        "type": "integer",
                        "example": "0"
                    }
                },
                "type": "object"
            },
            "PostUpdateRequest": {
                "required": [
                    "title",
                    "content"
                ],
                "properties": {
                    "title": {
                        "description": "Post title",
                        "type": "string",
                        "example": ""
                    },
                    "content": {
                        "description": "Post content",
                        "type": "string",
                        "example": ""
                    },
                    "status": {
                        "description": "Post status",
                        "type": "integer",
                        "example": "0"
                    }
                },
                "type": "object"
            },
...
        }
    }
}